+++ CHANGELOG.md
@@ -30,8 +30,9 @@
 - Remove deprecated fallback property of Locale entity
 - Add a generateProductCompletenesses method on CompletenessGeneratorInterface, to generate completeness for one product
 - Add setCompletenesses and getCompletenesses method on ProductInterface and Product class
-- Add methods applyFilterByField, applySorterByField and update methods applyFilterByAttribute, applySorterByAttribute in ProductRepositoryInterface
+- Add methods applyFilterByField, applySorterByField and update methods applyFilterByAttribute, applySorterByAttribute, deleteProductIds in ProductRepositoryInterface
 - Remove methods setLocale/getLocale, setScope/getScope, setConfiguration/getConfiguration from ProductRepositoryInterface
+- Remove methods setLocale/getLocale, setScope/getScope from ProductManager
 - Move findAllByAttributes and findOneByWithValues from FlexibleEntityRepositoryInterface to ProductRepositoryInterface
 - Move setFlexibleQueryBuilder, findAllByAttributes, findOneByWithValues, getFlexibleQueryBuilder, addJoinToValueTables, findAllByAttributesQB from FlexibleEntityRepository to ProductRepository (ORM)
 - Move FilterBundle/Filter/ScopeFilter.php, ProductCompletenessFilter.php, ProductGroupsFilter.php, CategoryFilter.php -> FilterBundle/Filter/Product/ScopeFilter.php, CompletenessFilter, GroupsFilter.php, CategoryFilter.php
@@ -53,7 +54,6 @@
 - Move mass export in specific controller
 - Add an affectsCompleteness method to MassEditActionInterface to indicate whether performing the mass action requires recalculating the product completeness
 - Remove DeleteMassActionHandler, replaced by ProductDeleteMassActionHandler
-- Add deleteProductIds in ProductRepository
 - Change product REST API data and url format
 - Remove incomplete REST API for getting multiple products
 - Remove Router dependency from json ProductNormalizer